refact(graphic): seperate view and model
